Final upload for HW2
I've been working on a private repo that I created so I have all of the files I've created throughout the production of this assignment. I can share the other repo with you if you'd like to see my workflow, commits, and branches. This should be everything to run the game though