Health and level indicators added
I increased the tower's base health to 300 which is double what it was. I also made the healing mechanic in the upgrade menu fully heal the tower. Both of these changes are designed to make the game significantly easier since playing it through I could barely reach level 3 which isn't near the hardest level. I also added level indicators hovering over the timers located in each level so the player knows when they've beaten the level.